Optimize symspell

+++ spellcheck/src/symspell/composition.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone)]
 pub struct Composition {
     pub segmented_string: String,
-    pub distance_sum: i64,
+    pub distance_sum: i32,
     pub prob_log_sum: f64,
 }
